+++ tests/data/forms/nested/form.u.js
@@ -37,6 +37,7 @@ class NewForm extends Form {
   options() {
     return {
       validateOnChange: true,
+      showErrorsOnChange: false,
     };
   }